added 'Select->Save selection as footprint" option
This menu option saves selected items into a footprint
file. It can save items inside 'elements' (ie. groups
of items), but the whole element must be also selected.
Selecting an individual item inside an unselected element
will not internally copy the item to a buffer and therefore
the item will not be saved to the footprint. Supporting
such feature would be possible, but requires more work.